Elizeo - Changelog 10-15-2024
Made a new script called CharacterUnit where it is the same as TestUnit, except that the stats uses Leland's CharactersBaseScript and his Scriptable Objects as stats.
^^
"As a player, I want to have my stats carry from battle to battle for consistency"

Also fixed some bugs with the HPslider using the SetHP function on the BattleHUD  script.